managed folders

+++ cleaner.sh
@@ -1,5 +1,5 @@
 for file in *; do
-    if [ "${file##*.}" != "tex" ] && [ "${file##*.}" != "bib" ] && [ "$file" != "images" ]&&[ "${file##*.}" != "sh" ]; then
+    if [ "${file##*.}" != "tex" ] && [ "${file##*.}" != "bib" ] && [ "$file" != "images" ]&&[ "${file##*.}" != "sh" ] &&[ "${main.}" != "pdf" ]; then
         rm "$file"
         echo "Deleted: $file"
     fi
